Check origin matches expected value
When a new key handle is registered with a u2f device, the origin is
encoded into the key handle. When a u2f device is authenticating it
verifies that the origin corresponds to the encoded origin in the
requested key handle, and puts the origin into the clientData in the
response.

To prevent a MITM who intermediated a registration we check that the
origin in the clientData corresponds with the expected value.