Add a public_name field to the u2f library stanza
This field is necessary to tell Dune to include this library as part of
`dune install`. Without this, the library isn't visible to users
installing `u2f` via e.g. `opam`.

A side-effect of this change is that `dune build @doc` will now generate
the documentation for this library correctly.